requirements.txt
streamlit
yfinance
pandas
requests
openai
matplotlib
textblob
schedule
python-dotenv
alpha_vantage
alpha_vantage
feedparser
beautifulsoup4
supabase
resend